MOCARNY PR - Wykrywanie rejestracji in real time
